[Remote Store] Added NRTReplicationNoOpEngine with NoOpTranslogManager
Replica Recovery is Working along with no-replication call for primary term validation
